+++ src/CharcFolder/Statue.py
@@ -1,0 +1,64 @@
+import Monster
+
+import src
+
+
+class Statue(Monster):
+    """
+    the class for animated statues
+    intended as temple guards
+    """
+
+    def __init__(
+        self,
+        display="@@",
+        xPosition=0,
+        yPosition=0,
+        quests=None,
+        automated=True,
+        name="Statue",
+        creator=None,
+        characterId=None,
+    ):
+        """
+        basic state setting
+
+        Parameters:
+            display: how the mouse should look like
+            xPosition: obsolete, ignore
+            yPosition: obsolete, ignore
+            quests: obsolete, ignore
+            automated: obsolete, ignore
+            name: obsolete, ignore
+            creator: obsolete, ignore
+            characterId: obsolete, ignore
+        """
+        if quests is None:
+            quests = []
+
+        super().__init__(
+            display,
+            xPosition,
+            yPosition,
+            quests,
+            automated,
+            name,
+            creator=creator,
+            characterId=characterId,
+        )
+        self.charType = "Statue"
+        self.specialDisplay = "@@"
+        self.baseDamage = 10
+
+    def changed(self, tag="default", info=None):
+        if tag == "pickup bolted fail":
+            info["item"].destroy()
+        super().changed(tag, info)
+
+    def die(self, reason=None, addCorpse=True):
+        """
+        die without leaving a corpse
+        """
+        super().die(reason, addCorpse=False)
+
+src.characters.add_character(Statue)
