+++ src/CharcFolder/Statuette.py
@@ -1,0 +1,49 @@
+import Monster
+
+import src
+
+
+class Statuette(Monster):
+    """
+    the class for a small statue
+    is intended as temple guard
+    """
+    def __init__(
+        self,
+        display="st",
+        xPosition=0,
+        yPosition=0,
+        quests=None,
+        automated=True,
+        name="Statuette",
+        creator=None,
+        characterId=None,
+    ):
+        """
+        basic state setting
+        """
+        if quests is None:
+            quests = []
+
+        super().__init__(
+            display,
+            xPosition,
+            yPosition,
+            quests,
+            automated,
+            name,
+            creator=creator,
+            characterId=characterId,
+        )
+        self.charType = "Statuette"
+        self.specialDisplay = "st"
+
+    def changed(self, tag="default", info=None):
+        if tag == "pickup bolted fail":
+            info["item"].destroy()
+        super().changed(tag, info)
+
+    def die(self, reason=None, addCorpse=True):
+        super().die(reason, addCorpse=False)
+
+src.characters.add_character(Statuette)
