+++ src/__init__.py
@@ -1,3 +1,4 @@
+import src.CharcFolder
 import src.items
 import src.itemFolder
 import src.questFolder
